LearnedDBComponentsLLM/requirements.txt
# ── Core ML ───────────────────────────────────────────────
torch
numpy
scipy
pandas
matplotlib

# ── Database ──────────────────────────────────────────────
psycopg2-binary

# ── SQL parsing ───────────────────────────────────────────
sqlglot

# ── LLM / Ollama ─────────────────────────────────────────
langchain
langchain-ollama
langgraph
requests

# ── Configuration ─────────────────────────────────────────
python-dotenv

# ── Data / Excel ──────────────────────────────────────────
openpyxl
